ui improvement, confirmation modal, pdf to brochure storage name

+++ src/components/cards/memberCard.svelte
@@ -15,6 +15,7 @@
     MenuItem,
   } from '@rgossiaux/svelte-headlessui';
   import { createEventDispatcher } from 'svelte';
+  import ConfirmationModal from '@comp/modals/confirmationModal.svelte';
 
   export let permissions;
   export let roles = [];
@@ -25,6 +26,10 @@
   let selectedRole = '';
   let memberRole = null;
   const dispatch = createEventDispatcher();
+  let showModal = false;
+  let roleID = null;
+  let roleName = null;
+  const toggleModal = () => (showModal = !showModal);
 
   const passDeletedMemberData = (member) =>
     dispatch('deletedMemberData', member);
@@ -113,12 +118,41 @@
     }
   };
 
+  let showDeleteMemberModal = false;
+  const toggleDeleteMemberModal = () =>
+    (showDeleteMemberModal = !showDeleteMemberModal);
+
   $: if (member.team_member_id) getMembersRole();
 </script>
 
+<ConfirmationModal
+  heading="Are you sure to change your role?"
+  buttonLabel="Yes, i am sure."
+  {showModal}
+  {toggleModal}
+  on:click={async () => {
+    await setMemberRole(roleID);
+    selectRole(roleName);
+    showModal = false;
+  }}
+/>
+
+<ConfirmationModal
+  isDelete
+  heading="Are you sure you want to delete"
+  text="yourself from team?"
+  buttonLabel="Delete"
+  showModal={showDeleteMemberModal}
+  toggleModal={toggleDeleteMemberModal}
+  on:click={async () => {
+    await deleteMemberHandler(member?.team_member_id?.id, member);
+    showDeleteMemberModal = false;
+  }}
+/>
+
 {#if member.team_member_id}
   {#if !permissions.readMembers}
-    {#if $user.id === member.team_member_id.uid}
+    {#if $user?.id === member.team_member_id.uid}
       <div class="flex flex-col justify-between">
         <div
           class="flex flex-col justify-between w-full h-full bg-neutral-800 rounded-md"
@@ -181,7 +215,7 @@
               <div>
                 <h2 class="text-neutral-300 text-xs mt-3">Card:</h2>
                 <p class="text-neutral-300 text-sm">
-                  {#if member?.type === 'pvc'}
+                  {#if member?.card_id?.type === 'pvc'}
                     PVC
                   {:else}
                     {member?.card_id?.type?.charAt(0).toUpperCase() +
@@ -195,18 +229,18 @@
           </div>
 
           <div
-            class="flex relative w-full justify-between items-center bg-neutral-900 rounded-b-md p-4"
+            class="flex relative h-20 w-full justify-between items-center bg-neutral-900 rounded-b-md p-4"
           >
-            <p
-              class="text-white border-2 border-neutral-700 flex justify-between items-center h-8 text-sm p-2 gap-2 rounded-md relative"
-            >
-              {selectedRole !== ''
-                ? selectedRole.charAt(0).toUpperCase() + selectedRole.slice(1)
-                : memberRole?.role_name
-                ? memberRole.role_name.charAt(0).toUpperCase() +
-                  memberRole.role_name.slice(1)
-                : 'No role'}
-            </p>
+            {#if permissions.readRoles}
+              <p
+                class="text-white border-2 border-neutral-700 flex justify-between items-center h-8 text-sm p-2 gap-2 rounded-md relative"
+              >
+                {memberRole?.role_name === 'superadmin'
+                  ? 'Super Admin'
+                  : memberRole?.role_name?.charAt(0).toUpperCase() +
+                    memberRole?.role_name?.slice(1)}
+              </p>
+            {/if}
 
             <!-- <SwitchButton
               on:change={async () => await setStatus()}
@@ -279,7 +313,7 @@
             <div>
               <h2 class="text-neutral-300 text-xs mt-3">Card:</h2>
               <p class="text-neutral-300 text-sm">
-                {#if member?.type === 'pvc'}
+                {#if member?.card_id?.type === 'pvc'}
                   PVC
                 {:else}
                   {member?.card_id?.type?.charAt(0).toUpperCase() +
@@ -318,11 +352,16 @@
                 class="top-0 right-0 z-50 relative mb-20 rounded-md flex flex-col bg-neutral-900 shadow-md border border-neutral-700 p-2 w-64"
               >
                 <MenuItem
-                  on:click={async () =>
-                    await deleteMemberHandler(
-                      member?.team_member_id?.id,
-                      member
-                    )}
+                  on:click={async () => {
+                    if ($user?.id === member.team_member_id.uid) {
+                      toggleDeleteMemberModal();
+                    } else {
+                      await deleteMemberHandler(
+                        member?.team_member_id?.id,
+                        member
+                      );
+                    }
+                  }}
                   class="flex hover:bg-neutral-800 text-red-600 px-2 py-2 rounded-md cursor-pointer"
                 >
                   Remove user
@@ -372,8 +411,14 @@
                     <MenuItem
                       class="flex hover:bg-neutral-700 px-2 py-2 rounded-md cursor-pointer"
                       on:click={async () => {
-                        await setMemberRole(item.id);
-                        selectRole(item.role_name);
+                        if ($user?.id === member.team_member_id.uid) {
+                          roleID = item.id;
+                          roleName = item.role_name;
+                          toggleModal();
+                        } else {
+                          await setMemberRole(item.id);
+                          selectRole(item.role_name);
+                        }
                       }}
                     >
                       {item.role_name.charAt(0).toUpperCase() +
@@ -383,8 +428,14 @@
                   <MenuItem
                     class="flex hover:bg-neutral-700 px-2 py-2 rounded-md cursor-pointer"
                     on:click={async () => {
-                      await setMemberRole(roleId[0].id);
-                      selectRole(roleId[0].name);
+                      if ($user?.id === member.team_member_id.uid) {
+                        roleID = roleId[0].id;
+                        roleName = roleId[0].name;
+                        toggleModal();
+                      } else {
+                        await setMemberRole(roleId[0].id);
+                        selectRole(roleId[0].name);
+                      }
                     }}
                   >
                     Super Admin
@@ -392,8 +443,14 @@
                   <MenuItem
                     class="flex hover:bg-neutral-700 px-2 py-2 rounded-md cursor-pointer"
                     on:click={async () => {
-                      await setMemberRole(roleId[1].id);
-                      selectRole(roleId[1].name);
+                      if ($user?.id === member.team_member_id.uid) {
+                        roleID = roleId[1].id;
+                        roleName = roleId[1].name;
+                        toggleModal();
+                      } else {
+                        await setMemberRole(roleId[1].id);
+                        selectRole(roleId[1].name);
+                      }
                     }}
                   >
                     Member
@@ -411,7 +468,7 @@
       </div>
     </div>
   {/if}
-{:else}
+{:else if permissions.readMembers}
   <div class="flex flex-col justify-between">
     <div
       class="flex flex-col justify-between w-full h-full bg-neutral-800 rounded-md"
@@ -449,7 +506,7 @@
       </div>
 
       <div
-        class="flex relative w-full h-full justify-between items-center bg-neutral-900 rounded-b-md p-4"
+        class="flex relative w-full h-20 justify-between items-center bg-neutral-900 rounded-b-md p-4"
       />
     </div>
   </div>
